requirements.txt
# Core dependencies
websockets==12.0
aiohttp==3.9.1
asyncio==3.4.3
python-dotenv==1.0.0

# AI and API dependencies
openai==1.3.7
anthropic==0.7.8

# Data processing
pandas==2.2.0
numpy==1.26.4

# Database
pymongo==4.6.0

# HTTP requests
requests==2.31.0

# Logging
loguru==0.7.2

# Web scraping and text processing
beautifulsoup4==4.12.2
textblob==0.17.1

# Utilities
tenacity==8.2.3
schedule==1.2.0 